% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectFunctions.R
\name{createFullProject}
\alias{createFullProject}
\title{createFullProject:
THE MOST IMPORTANT FUNCTION}
\usage{
createFullProject(name)
}
\arguments{
\item{name}{A character string}
}
\value{
An entire project directory structure
}
\description{
createFullProject:
THE MOST IMPORTANT FUNCTION
}
\examples{
crearteFullProject("My first project")
}
