% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectFunctions.R
\name{fileMasterCreate}
\alias{fileMasterCreate}
\title{fileMasterCreate:
create a .R script for sourcing example}
\usage{
fileMasterCreate(name)
}
\arguments{
\item{name}{A character string}
}
\value{
a master .Rmd file for sourcing scripts
}
\description{
fileMasterCreate:
create a .R script for sourcing example
}
\examples{
fileMasterCreate("My first project")
}
