% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectFunctions.R
\docType{package}
\name{fileStruct}
\alias{fileStruct}
\title{fileStruct: A package for developing reproducable project structures}
\description{
The fileStruct package provides one main function, with some additional "helper"
functions for creating an organized workflows.
}
\section{The fileStruct() function}{

createFullProject() is the most important function here, and creates a file structure
based on the following link: https://julianreif.com/guide/

Simply set a directory and run the function with your desired project name to
create a clean and efficient project structure.
}

