% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectFunctions.R
\name{projectCreate}
\alias{projectCreate}
\title{projectCreate:}
\usage{
projectCreate(name)
}
\arguments{
\item{name}{A character string}
}
\value{
an r.proj of name "name"
}
\description{
projectCreate:
}
\examples{
projectCreate("My first project")
}
