% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectFunctions.R
\name{scriptCreate}
\alias{scriptCreate}
\title{scriptCreate:}
\usage{
scriptCreate(name)
}
\arguments{
\item{name}{A character string}
}
\value{
a simple .R script in analysis/scripts
}
\description{
scriptCreate:
}
\examples{
scriptCreate("My first project")
}
