% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectFunctions.R
\name{createSubdirectories}
\alias{createSubdirectories}
\title{createSubdirectories:}
\usage{
createSubdirectories(name)
}
\arguments{
\item{name}{A character string}
}
\value{
a directory structure based on https://julianreif.com/guide/
}
\description{
createSubdirectories:
}
\examples{
createSubdirectories("My first project")
}
